+++ railties/CHANGELOG.md
@@ -1,3 +1,10 @@
+*   Set `dump_schema_after_migration` config values in production.
+
+    Set `config.active_record.dump_schema_after_migration` as false
+    in the generated `config/environments/production.rb` file.
+
+    *Emil Soman*
+
 *   Added Thor-action for creation of migrations.
 
     Fixes #13588 and #12674.